requirements.txt
pandas==1.3.3
scikit-learn==0.24.2
python-dotenv==0.19.1